+++ src/qtgui/dockfft.cpp
@@ -600,7 +600,7 @@ void DockFft::setWfResolution(quint64 msec_per_line)
 {
     float res = 1.0e-3f * (float)msec_per_line;
 
-    ui->wfResLabel->setText(QString("Res: %1 s").arg(res, 0, 'f', 2));
+    ui->wfResLabel->setText(QString("Res: %1 s").arg((double)res, 0, 'f', 2));
 }
 
 /**
@@ -777,11 +777,11 @@ void DockFft::updateInfoLabels(void)
 
     rbw = m_sample_rate / size;
     if (rbw < 1.e3f)
-        ui->fftRbwLabel->setText(QString("RBW: %1 Hz").arg(rbw, 0, 'f', 1));
+        ui->fftRbwLabel->setText(QString("RBW: %1 Hz").arg((double)rbw, 0, 'f', 1));
     else if (rbw < 1.e6f)
-        ui->fftRbwLabel->setText(QString("RBW: %1 kHz").arg(1.e-3f * rbw, 0, 'f', 1));
+        ui->fftRbwLabel->setText(QString("RBW: %1 kHz").arg(1.e-3 * (double)rbw, 0, 'f', 1));
     else
-        ui->fftRbwLabel->setText(QString("RBW: %1 MHz").arg(1.e-6f * rbw, 0, 'f', 1));
+        ui->fftRbwLabel->setText(QString("RBW: %1 MHz").arg(1.e-6 * (double)rbw, 0, 'f', 1));
 
     rate = fftRate();
     if (rate == 0)
@@ -795,5 +795,5 @@ void DockFft::updateInfoLabels(void)
         else
             ovr = 100 * (1.f - interval_samples / size);
     }
-    ui->fftOvrLabel->setText(QString("Overlap: %1%").arg(ovr, 0, 'f', 0));
+    ui->fftOvrLabel->setText(QString("Overlap: %1%").arg((double)ovr, 0, 'f', 0));
 }
